update:qrScanのバリデーションの分離
