+++ lib/Widget/qrScan.dart
@@ -2,8 +2,7 @@ import 'dart:convert';
 
 import 'package:barcode_scan/barcode_scan.dart';
 import 'package:flutter/services.dart';
-import 'package:intl/intl.dart';
-import 'package:stamp_app/Util/enumCheckString.dart';
+import 'package:stamp_app/Util/Validation.dart';
 
 // 読み込み結果を表示するダイアログ
 Future<ScanResult> qrScan() async {
@@ -37,26 +36,5 @@ Future<ScanResult> qrScan() async {
     );
     result.rawContent = 'エラー: ${e.message}';
     return result;
-  }
-}
-
-// 読み込んだQRを検証するクラス
-class Validation {
-  static final String stampCheckString = CheckString.ok.checkStringValue;
-  static bool dateCheck(datetime) {
-    DateTime verificationDate =
-        DateFormat("yyyy/MM/dd HH:mm:ss").parseStrict(datetime);
-    DateTime now = DateTime.now();
-
-    if (now.isBefore(verificationDate)) return false;
-    return true;
-  }
-
-  static bool strCheck(String content) {
-    // 記号・半角を含まない
-    if (RegExp(r'^(?=.*[!-/:-@[-`{-~]).*$').hasMatch(content)) return false;
-    if (!content.contains(stampCheckString)) return false;
-    if (content.contains('http')) return false;
-    return true;
   }
 }
